+++ spec/honeykiq/periodic_reporter_spec.rb
@@ -1,4 +1,3 @@
-require 'honeykiq'
 require 'libhoney'
 
 RSpec.describe Honeykiq::PeriodicReporter do