+++ spec/spec_helper.rb
@@ -1,4 +1,5 @@
 require 'honeykiq'
+require 'pry'
 
 RSpec.configure do |config|
   Kernel.srand config.seed